refactor: split frontend common UI bundle v5
