+++ docs/dev-logs/2026-06-06-frontend-common-ui-split-bundle-v5.md
@@ -1,0 +1,15 @@
+# Frontend Common UI Split Bundle v5
+
+## Summary
+- Split `AssignmentCheckPageSections` into header, filters, and list sections.
+- Split `LectureStudioEditorFormSections` into basic info, operations, and assessment sections.
+- Split `CourseCreateCard` summary rendering into a dedicated section export.
+- Kept `MyShortformsPageSections` as a lightweight re-export barrel for the new subcomponents.
+
+## Verification
+- `npm run build:frontend`
+- `npm run test:backend`
+
+## Notes
+- The latest frontend build passed after the split.
+- Backend test suite passed unchanged at 74 tests.
